# Swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Sistema de Subastas API",
      "description": "API REST del Sistema de Subastas — UADE Desarrollo de Aplicaciones I, 1C 2026. Gestiona registración de postores, catálogo, subastas, pujas, ventas, multas, seguros y solicitudes de incorporación de bienes.",
      "contact": {
        "name": "Equipo DA1",
        "email": "agustin.torres@majily.com"
      },
      "version": "0.1.0"
    },
    "paths": {
...
      "/usuarios/{usuario_id}/medios-pago": {
        "post": {
          "tags": [
            "Medios de pago"
          ],
          "summary": "Registrar medio de pago para el usuario",
          "operationId": "crear_medio_pago_usuarios__usuario_id__medios_pago_post",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoCreate"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MedioPagoOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "Medios de pago"
          ],
          "summary": "Listar medios de pago del usuario",
          "operationId": "listar_medios_pago_usuarios__usuario_id__medios_pago_get",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MedioPagoOut"
                    },
                    "title": "Response Listar Medios Pago Usuarios  Usuario Id  Medios Pago Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/usuarios/{usuario_id}/medios-pago/{mp_id}/verificar": {
        "post": {
          "tags": [
            "Medios de pago"
          ],
          "summary": "Verificar (o rechazar) un medio de pago antes del inicio de la subasta",
          "operationId": "verificar_medio_pago_usuarios__usuario_id__medios_pago__mp_id__verificar_post",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            },
            {
              "name": "mp_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Mp Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoVerificar"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MedioPagoOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/usuarios/{usuario_id}/medios-pago/{mp_id}": {
        "delete": {
          "tags": [
            "Medios de pago"
          ],
          "summary": "Eliminar medio de pago",
          "operationId": "eliminar_medio_pago_usuarios__usuario_id__medios_pago__mp_id__delete",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            },
            {
              "name": "mp_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Mp Id"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successful Response"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/articulos": {
        "get": {
          "tags": [
            "Artículos"
          ],
          "summary": "Listar Articulos",
          "operationId": "listar_articulos_articulos_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/ArticuloOut"
                    },
                    "type": "array",
                    "title": "Response Listar Articulos Articulos Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Artículos"
          ],
          "summary": "Crear Articulo",
          "operationId": "crear_articulo_articulos_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticuloCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ArticuloOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/articulos/{articulo_id}": {
        "get": {
          "tags": [
            "Artículos"
          ],
          "summary": "Obtener Articulo",
          "operationId": "obtener_articulo_articulos__articulo_id__get",
          "parameters": [
            {
              "name": "articulo_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Articulo Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ArticuloOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/articulos/{articulo_id}/deposito/{deposito_id}": {
        "put": {
          "tags": [
            "Artículos"
          ],
          "summary": "Asignar Deposito",
          "operationId": "asignar_deposito_articulos__articulo_id__deposito__deposito_id__put",
          "parameters": [
            {
              "name": "articulo_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Articulo Id"
              }
            },
            {
              "name": "deposito_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Deposito Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ArticuloOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/depositos": {
        "get": {
          "tags": [
            "Depósitos"
          ],
          "summary": "Listar Depositos",
          "operationId": "listar_depositos_depositos_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/DepositoOut"
                    },
                    "type": "array",
                    "title": "Response Listar Depositos Depositos Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Depósitos"
          ],
          "summary": "Crear Deposito",
          "operationId": "crear_deposito_depositos_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositoCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DepositoOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/seguros": {
        "get": {
          "tags": [
            "Seguros"
          ],
          "summary": "Listar Seguros",
          "operationId": "listar_seguros_seguros_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/SeguroOut"
                    },
                    "type": "array",
                    "title": "Response Listar Seguros Seguros Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Seguros"
          ],
          "summary": "Crear Seguro",
          "operationId": "crear_seguro_seguros_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeguroCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SeguroOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "ArticuloCreate": {
          "properties": {
            "numero_pieza": {
              "type": "string",
              "title": "Numero Pieza"
            },
            "descripcion": {
              "type": "string",
              "title": "Descripcion"
            },
            "precio_base": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ],
              "title": "Precio Base"
            },
            "moneda": {
              "type": "string",
              "title": "Moneda",
              "default": "ARS"
            },
            "dueno_actual_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dueno Actual Id"
            },
            "artista": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artista"
            },
            "fecha_obra": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Obra"
            },
            "historia": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Historia"
            },
            "cantidad_elementos": {
              "type": "integer",
              "title": "Cantidad Elementos",
              "default": 1
            },
            "imagenes": {
              "items": {
                "$ref": "#/components/schemas/ImagenArticuloCreate"
              },
              "type": "array",
              "title": "Imagenes"
            }
          },
          "type": "object",
          "required": [
            "numero_pieza",
            "descripcion",
            "precio_base"
          ],
          "title": "ArticuloCreate"
        },
        "ArticuloOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "numero_pieza": {
              "type": "string",
              "title": "Numero Pieza"
            },
            "descripcion": {
              "type": "string",
              "title": "Descripcion"
            },
            "precio_base": {
              "type": "string",
              "title": "Precio Base"
            },
            "moneda": {
              "type": "string",
              "title": "Moneda"
            },
            "dueno_actual_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dueno Actual Id"
            },
            "artista": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artista"
            },
            "fecha_obra": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Obra"
            },
            "historia": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Historia"
            },
            "cantidad_elementos": {
              "type": "integer",
              "title": "Cantidad Elementos"
            },
            "estado": {
              "$ref": "#/components/schemas/EstadoArticulo"
            },
            "deposito_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deposito Id"
            },
            "imagenes": {
              "items": {
                "$ref": "#/components/schemas/ImagenArticuloOut"
              },
              "type": "array",
              "title": "Imagenes",
              "default": []
            }
          },
          "type": "object",
          "required": [
            "id",
            "numero_pieza",
            "descripcion",
            "precio_base",
            "moneda",
            "cantidad_elementos",
            "estado"
          ],
          "title": "ArticuloOut"
        },
...
        "DepositoCreate": {
          "properties": {
            "nombre": {
              "type": "string",
              "title": "Nombre"
            },
            "direccion": {
              "type": "string",
              "title": "Direccion"
            },
            "ciudad": {
              "type": "string",
              "title": "Ciudad"
            }
          },
          "type": "object",
          "required": [
            "nombre",
            "direccion",
            "ciudad"
          ],
          "title": "DepositoCreate"
        },
        "DepositoOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "nombre": {
              "type": "string",
              "title": "Nombre"
            },
            "direccion": {
              "type": "string",
              "title": "Direccion"
            },
            "ciudad": {
              "type": "string",
              "title": "Ciudad"
            }
          },
          "type": "object",
          "required": [
            "id",
            "nombre",
            "direccion",
            "ciudad"
          ],
          "title": "DepositoOut"
        },
        "EstadoArticulo": {
          "type": "string",
          "enum": [
            "disponible",
            "en_subasta",
            "vendido",
            "retirado"
          ],
          "title": "EstadoArticulo"
        },
        "EstadoMedioPago": {
          "type": "string",
          "enum": [
            "pendiente",
            "verificado",
            "rechazado"
          ],
          "title": "EstadoMedioPago"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "ImagenArticuloCreate": {
          "properties": {
            "url": {
              "type": "string",
              "title": "Url"
            },
            "orden": {
              "type": "integer",
              "title": "Orden",
              "default": 0
            }
          },
          "type": "object",
          "required": [
            "url"
          ],
          "title": "ImagenArticuloCreate"
        },
        "ImagenArticuloOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "url": {
              "type": "string",
              "title": "Url"
            },
            "orden": {
              "type": "integer",
              "title": "Orden"
            }
          },
          "type": "object",
          "required": [
            "id",
            "url",
            "orden"
          ],
          "title": "ImagenArticuloOut"
        },
...
        "MedioPagoCreate": {
          "properties": {
            "tipo": {
              "$ref": "#/components/schemas/TipoMedioPago"
            },
            "titular": {
              "type": "string",
              "title": "Titular"
            },
            "detalle": {
              "type": "string",
              "title": "Detalle"
            },
            "pais": {
              "type": "string",
              "title": "Pais"
            },
            "monto_garantia": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Monto Garantia"
            },
            "moneda": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Moneda"
            }
          },
          "type": "object",
          "required": [
            "tipo",
            "titular",
            "detalle",
            "pais"
          ],
          "title": "MedioPagoCreate"
        },
        "MedioPagoOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "usuario_id": {
              "type": "integer",
              "title": "Usuario Id"
            },
            "tipo": {
              "$ref": "#/components/schemas/TipoMedioPago"
            },
            "titular": {
              "type": "string",
              "title": "Titular"
            },
            "detalle": {
              "type": "string",
              "title": "Detalle"
            },
            "pais": {
              "type": "string",
              "title": "Pais"
            },
            "monto_garantia": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Monto Garantia"
            },
            "moneda": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Moneda"
            },
            "estado": {
              "$ref": "#/components/schemas/EstadoMedioPago"
            },
            "verificado": {
              "type": "boolean",
              "title": "Verificado"
            }
          },
          "type": "object",
          "required": [
            "id",
            "usuario_id",
            "tipo",
            "titular",
            "detalle",
            "pais",
            "estado",
            "verificado"
          ],
          "title": "MedioPagoOut"
        },
        "MedioPagoVerificar": {
          "properties": {
            "verificado": {
              "type": "boolean",
              "title": "Verificado"
            }
          },
          "type": "object",
          "required": [
            "verificado"
          ],
          "title": "MedioPagoVerificar"
        },
...
        "SeguroCreate": {
          "properties": {
            "nro_poliza": {
              "type": "string",
              "title": "Nro Poliza"
            },
            "compania": {
              "type": "string",
              "title": "Compania"
            },
            "beneficiario_id": {
              "type": "integer",
              "title": "Beneficiario Id"
            },
            "monto_cubierto": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ],
              "title": "Monto Cubierto"
            },
            "moneda": {
              "type": "string",
              "title": "Moneda",
              "default": "ARS"
            },
            "articulo_ids": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Articulo Ids"
            }
          },
          "type": "object",
          "required": [
            "nro_poliza",
            "compania",
            "beneficiario_id",
            "monto_cubierto",
            "articulo_ids"
          ],
          "title": "SeguroCreate"
        },
        "SeguroOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "nro_poliza": {
              "type": "string",
              "title": "Nro Poliza"
            },
            "compania": {
              "type": "string",
              "title": "Compania"
            },
            "beneficiario_id": {
              "type": "integer",
              "title": "Beneficiario Id"
            },
            "monto_cubierto": {
              "type": "string",
              "title": "Monto Cubierto"
            },
            "moneda": {
              "type": "string",
              "title": "Moneda"
            }
          },
          "type": "object",
          "required": [
            "id",
            "nro_poliza",
            "compania",
            "beneficiario_id",
            "monto_cubierto",
            "moneda"
          ],
          "title": "SeguroOut"
        },
...
        "TipoMedioPago": {
          "type": "string",
          "enum": [
            "cuenta_bancaria",
            "tarjeta_credito",
            "cheque_certificado"
          ],
          "title": "TipoMedioPago"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        },
...
      }
    }
  }